nanopore-pipeline/pyproject.toml
[project]
name = "nanopore-pipeline"
version = "0.1.0"
description = "Dagster pipeline for downloading and processing nanopore methylation data"
readme = "README.md"
authors = [
    { name = "antonkulaga", email = "antonkulaga@gmail.com" }
]
requires-python = ">=3.11,<3.14"
dependencies = [
    "just-biomarkers>=0.1.0",
    "dagster>=1.12.19",
    "dagster-webserver>=1.12.19",
    "dagster-polars>=0.27.9",
    "python-dotenv>=1.2.2",
    "typer>=0.24.1",
    "psutil>=7.2.2",
    "requests>=2.32.0",
    "beautifulsoup4>=4.13.0",
    "platformdirs>=4.0.0",
    "rich>=14.3.3",
]

[project.scripts]
nanopore-pipeline = "nanopore_pipeline.definitions:main"
nanopore = "nanopore_pipeline.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nanopore_pipeline"]

[tool.uv.sources]
just-biomarkers = { workspace = true }
